finished functionality for controllers contacts.js and vehicles.js
